Update relative imports due to file relocation
