+++ packages/api/portfolioDrafts/createPortfolioDraft.ts
@@ -1,11 +1,11 @@
 import { PutCommand } from "@aws-sdk/lib-dynamodb";
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import { v4 as uuid } from "uuid";
-import { ErrorCodes } from "./models/Error";
-import { PortfolioDraftSummary } from "./models/PortfolioDraftSummary";
-import { ProvisioningStatus } from "./models/ProvisioningStatus";
-import { dynamodbClient as client } from "./utils/dynamodb";
-import { ApiSuccessResponse, ErrorResponse, ErrorStatusCode, SuccessStatusCode } from "./utils/response";
+import { ErrorCodes } from "../models/Error";
+import { PortfolioDraftSummary } from "../models/PortfolioDraftSummary";
+import { ProvisioningStatus } from "../models/ProvisioningStatus";
+import { dynamodbClient as client } from "../utils/dynamodb";
+import { ApiSuccessResponse, ErrorResponse, ErrorStatusCode, SuccessStatusCode } from "../utils/response";
 
 const TABLE_NAME = process.env.ATAT_TABLE_NAME;
 // const PRIMARY_KEY = process.env.PRIMARY_KEY || "";
